Rename splitter to plist
